Update How To Section
Add in links to notebooks as the Earthdata Cloud Cookbook How To section is still a work in progress. This can be modified later. Note: The .md files are linked externally still because we need to update the import code to work with importing .md files into the podaac repo, not just .ipynb files.