Update CMR portion
I updated the CMR portion of the notebook to access the two specific files after obtaining a list of them rather than calling them out individually and then also combined the mosaic portion into a function to make it more concise per timestamp. I also fixed the pathnames to match with what we currently have in the github directory.
Updated the gitignore file to ignore the datasets/data_downloads folder.