Fixed errors in code
I found that using a PC instead of a mac made it so I had to modify the code a tad from Jack's original code. For instance, I needed to specify .astype(np.int64) instead of just .astype(int) for the quality flag values to show up correctly. Also I needed to .load() in the qual variable in the first instance of use.